Make sure HWP ammo does not show up in "craft"equipment screens.
Downside is that you cannot show bigSprites in the ufopaedia for HWP-ammo.
Luckily that is currently not used for HWP-ammo.

Additionally removed "ignoreInBaseDefense" variables. Those are not needed if an item has no bigSprite.
Those tags only influence the availability of an item on the battlescape, it does **not** affect equipment screens